use other version of secure send
